+++ example/month_example.dart
@@ -1,6 +1,6 @@
 import 'package:grizzly_range/grizzly_range.dart';
 
-main() {
+void main() {
   print(MonthRange(DateTime(2020, 2, 29), DateTime(2032, 2, 30), 6).toList());
   print(MonthRange(DateTime(2019, 1, 31), DateTime(2022, 1, 31), 1).toList());