+++ example/time_example.dart
@@ -1,6 +1,6 @@
 import 'package:grizzly_range/grizzly_range.dart';
 
-main() {
+void main() {
   print(TimeRange(
       DateTime(2019, 1, 1), DateTime(2019, 1, 20), Duration(days: 1)));
   print(TimeRange(